refactor(racing-car): InputValidation 클래스 view 패키지에 추가
