+++ src/main/java/racingcar/view/InputValidation.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 import racingcar.UserInputException;
